Updated mediawiki/extensions Project: mediawiki/extensions/BounceHandler  c3d6b17b6cddf366ad321016f732a6aa6000e1fb
Notify administrators on API failing to parse bounce emails

A bounce email, can go un-processed, even after it passed valid
VERP checks. The wiki admin needs to be notified on this event,
so that he can improve his regex functions. Added a test to check
existence of $user in BounceHandlerHooks

Change-Id: I1c3577a9ed87f6a2656d20e319d9f66dd38d3d70