Updated mediawiki/extensions Project: mediawiki/extensions/Math  a25078ff9d62ba75ff9ba78eb149a03586154556
Remove unused empty production for `litsq_zq` in texvccheck

It is impossible to take the empty alternative for this production,
since any input which matches `litsq_aq` would instead match `lit_aq`
if the empty alternative is used.  In addition, there is no action
clause for the empty alternative, so it wouldn't typecheck if the
empty alternative was possible.

This has no effect on the generated parser, but clarifies the grammar
(and makes ports easier!).

Change-Id: I32c85d2c74ab047b4b52401248d5dc7fc3201705