Updated mediawiki/extensions Project: mediawiki/extensions/Flow  b282cc3c034db34632351035dc615edc164f8594
Convert textareas into editor objects

This adds auto-expanding text-areas[1]

Basically, all textareas trigger an "editor" object to be created, which
abstracts editor functionality out of the forms: content is fetched via
mw.flow.editor.getContent( $node ).
This is important because we'll have other editors some day (like VE), which may
not just have this one textarea with the text in the format we want (textarea)

This editor code existed in our previous frontend codebase - we may want to make
changes for it to better match our existing codebase.

1: https://trello.com/c/lGdQUgv7/396-text-area-for-post-used-to-auto-expand-as-you-typed-more-post-rewrite-parity

Change-Id: I1ce819744de7124c0474ea19cb19c45688db401e