Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  dace5ef4876b360c09e0b72e5ccd565bf8997284
Drop dead experimental code

Warning: I honestly don't know what the purpose of this
UnitTestsList hook code was. But it's obvious that it's dead code
now, right? The array was emptied in previous commits, e.g. Iece4a23.
Should be easy to recover these snippets from the history if they
are needed again.

Change-Id: Iab6bf24c9602c63310c94d7e77d7975be07c7058