Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  0270c7eca5b9963048955c2e8dc5c41abf79cf3f
Add UnDeserializableValue to DataModelSerializationRoundtripTest

This roundtrip test failed because both the new and the legacy
unserializer (in Lib) checked the hash. This can't work if the
hash was (partly) created by the UnDeserializableValue class.

All hash ckecks are now removed from the new serializer.

This removes the last remaining check from the old serializer. See
https://github.com/wmde/WikibaseDataModelSerialization/commit/f993c186

Bug: 66839
Bug: 68565
Change-Id: I002e1024275e2270c3de94eda4b66aee792e00fb