Updated mediawiki/extensions Project: mediawiki/extensions/MwEmbedSupport  313e5c34d3dc723872870276d25222b0fda0b305
Do not specify any z-index for the play button

Originally this had a z-index of 1. This meant that the button
was unclickable in the pop up player since the parent had a
way higher z-index. Then it was changed in 
3bc774a6ccb30ac691225b256cc3189ca170bb08 to 1501, but that made
the play button from other media overlap the video player when
playing a video file, and caused it to overlap the player controls.

Removing the z-index declaration (making it be auto) causes the
button to be clickable in the pop up player, but not overlap
everything in the page, which is what we want.

Bug: 68479
Change-Id: Ic6b68f41dc5bbd3c16656fe7c83bf48a7a2deaba