Updated mediawiki/extensions Project: mediawiki/extensions/WikimediaIncubator  6461058dfe49659d3486be90f72d33b9ff15121c
Merge infopage and incubator i18n

This change merges the "incubator" and "infopage" directories
(including translations and credits) and moves the result in the
root i18n directory.

The merge was performed as there is no reason for there to be
multiple i18n files for the same language. Merging them
centralizes the translations in one location instead of two.
It's simpler.

(Tedious manual work is done when I'm fatally bored)

Change-Id: I524fb415111904377d0f6226683e4a837ff5338c