Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  2e8aec1d197ebd25bb4e2dc4f9e0e70b3f90b0b5
Reduce GUI jumping around when clicking edit

When clicking [edit] on a statement there is a graphical glitch
visible for a short time (while waiting for the API request).
The lines [save|remove|cancel] and [add qualifier] overlap. This
basic min-height avoids this overlapping and reduces the elements
jumping around (but can't completely avoid the jumping because it
can't know the exact height of the editor).

Change-Id: Ia5fa21b00f7d04b92dd5e90aa1ff985e97ac2765