Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  34e54b49b0689edb518b4e64c345fac627d5fdb7
Add order test to DataModelSerializationRoundtripTest

I'm not very happy with this test. It's very different from the
other tests because it relies on knowledge how the serialized arrays
are constructed. But I think it's good enough. Please keep in mind
that this test is temporary anyway. It will die with the old
serializers.

Change-Id: If0fc89c20b29104438aa28f40ddf5cebe60eb881