Updated mediawiki/extensions Project: mediawiki/extensions/Flow  76bbc4e5397b7a55399568050c5c372ad885dcd5
No-JS: Make "Start a new topic" a link to another page instead of open text fields

This patch will add CSS classes flow-js & flow-nojs. Those will be shown/hidden
according to client-js or client-nojs's presence.
Those will be used to either display the link to the seperate new-topic form
(no-js) or immediately display the form (js)

A new template has been added for the no-js new-topic form page.

The most notable change here is probably that the code to set the page title has
been taken out of View.php and moved to the specific blocks. This makes it
easier to set specific titles per action (which was requested for this feature,
and judging from a lingering @todo in the code, was already in the back of our
heads)

Change-Id: I809c2ae9d1337bb3081f652a158072c15470d488