Updated mediawiki/extensions Project: mediawiki/extensions/BounceHandler  b73c715c61bcf4dc1f4190b033ac870e5c7defbf
Fixed various issues with the BounceHandler extension

* The getUserDetails() used to return nothing earlier in the case
of a failed hash validation. Now the calling function checks deep
into the array to find whether the hash validation was successfull.
* Moved processEmail() to the parent class.
* The unsubscribe action used to lookup in the database according to
the email address, which can be faked. This has been changed to a
email_id AND user_id check to increase security.
* Fixed style issues

Bug: 69099
Change-Id: I234c72bfa465d2aa2cc36972d781b8f9f27ee3ec