Updated mediawiki/extensions Project: mediawiki/extensions/Translate  c64c1cb58a8aa9c00f9d6470f1b8bf19a6cec354
Skip source units on pressing tab key at Special:PageMigration

The source units are skipped and pressing the tab key while the
cursor is in a translation textarea takes the cursor to the next
translation textarea. This is achieved by setting the tabindex
attribute to -1 for all the source unit textarea's.

Bug: 69307
Change-Id: I098bb4c1e2e6a7140010f549057bd9118d5f79ae