Updated mediawiki/extensions Project: mediawiki/extensions/BounceHandler  d325bb411370c7aca4f3d5e1fef2e2efc0a1fd50
Improved the VERP generation by cutting down the hmac hash

The $local_part of a standard email should be < 64 which was not
taken care of earlier, and there was high chances of crossing it.
Cutting down securely the hmac hash looks fine, as per security
experts.
Added a prefix such as 'wiki-' so that the email-id looks good, and
easy to parse by exim.
Added documentation.

Change-Id: I167c018be679f6f1535824e94d5aa380dda069a7