Updated mediawiki/extensions Project: mediawiki/extensions/Flow  324f91bc5f4c7ace41fce22626a5cff732c73060
Remove redundant "Given I am on Flow page" steps

Now these three Scenarios share a "Given I am logged in
And I have created a Flow topic" Background, the last two don't need a
"Given I am on Flow page" that re-requests the current page.  It takes
time and led to a browser test failure where the login state was lost
(Firefox-sauce/146), so unless we actually want to check the menu is
there on reload, we can remove the step.

Tests still pass on ee-flow in Firefox with this change.

Change-Id: Icefa190526d02419c1cbffb70655890bdefc8773