Updated mediawiki/extensions Project: mediawiki/extensions/ContentTranslation  625006ec52c61ff511cfb296b4754603877f5bc6
Make the MTAbuse card sticky

Keep it in tools column till the MT percentage is below threshold.
We were keeping it in tools column, but it was not cooexisting with
other cards. Made it sticky.

Also, change the event trigger for the card. Now it listen for mw.cx.progress

Change-Id: Ic073634f6316e14221544383fad3bab87293f7d7