Updated mediawiki/extensions Project: mediawiki/extensions/TimedMediaHandler  bc08e2f3200c2130f1c63635ed9643ba90f4d2b5
Fix horribly broken way TMH was generating <track> urls

This just fixes the mechanism it was using to generate the urls.
The issue with the namespace missing is still outstanding.

Previously it used the wrong type of url encoding, and tried to
use ?action=raw from a short url entry point. Thus making totally
invalid urls.

Bug: 69453
Change-Id: I327ad84de8b6d16f8d17376add97689bb276d11a