Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  0c371f904cd8de07108a389e5921d5baeb95c429
QA: Fix login tests

Reset password link shows up if you are not logged in, so let's not log
in. Remove test checking for the message saying that the user is already
logged in. The new behavior in core simply redirects you to main page if
you are logged in and try to access Special:UserLogin.

Change-Id: I3b75cd879f1d81f1bd61dc0a6924f11d647671b8