Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  a7ff54edfbf5d62b251ac53bbff1846ceb62a297
Skip highlighting we don't need

Slightly changes behavior: only displays redirect or section match if
there isn't a title match.  This might not be perfect but it is in line
with our goals to reduce IO usage.

DEPLOYMENT: This does nothing until the upgrade to Elasticsearch 1.3
and experimental-highlight 0.0.11 both of which are scheduled soon.

Change-Id: I58f031adedf875e0f2723c5c21c7ca91e1088893