Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  a73a16ba47c695050d2dc31e7320b54be8240b58
Remove unused dependency from pagesuggester

This is a follow-up to I44a2240 and later patches in the same
chain. I'm aware some of these changes are a matter of taste.
This patch is mainly meant as a collection of suggestions.

I'm pretty sure about the dependency but it would be good if you
could check this.

Change-Id: I5cca87db5a8e06ffe70697c0d21ada01b1f6c595