Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  792cc65dbada1b34d3359f76f7fff1412e8dc01f
Refactor claims code out of EntityView

All code related to claims is moved into a separate ClaimsView which only
knows about a list of claims. This helps us getting rid of EntityView.

Note that lots of test code is copied from EntityViewTest which however
will be deleted once we remove EntityView so that there is no duplicate
code in the long term. Furthermore, the tests do not actually test
anything further than that the output is actually HTML. The ClaimsViewTest
can also be improved in another patch as this one would grow too big.

A problem at the moment is the ClaimHtmlGenerator which also accepts
statements which extend claims at the moment. However, when this will
change we may think of renaming this to StatementsView and only accept
statements. Or we find another way of using this to display both claims
and statements.

In future ClaimsView should perhaps only be created in ItemView and be
injected via the constructor. However, this can be done in another patch
as this one would get too big. To actually kill EntityView we have to
modify the Content classes (EntityContent, ItemContent and
PropertyContent) atm..

Change-Id: Iee15ec58f69abac9fb6a85c23d6196a309d1a651