Updated mediawiki/extensions Project: mediawiki/extensions/Math  403b238e1c503110bb49a6ea49df95f1aab3dc7b
Indicate that a PNG fallback can also be used for the MathML mode

The fallback image for the MathML mode is in SVG format, except for Internet
Explorer < 9 for which PNG is used instead. This commit makes the label of the
MathML mode sligthly more accurate, by mentioning that PNG may be used.

Change-Id: I77c8a8dcb7d2f111f029e1cd5b6f3442640be57e