Updated mediawiki/extensions Project: mediawiki/extensions/MultimediaViewer  d6c142fd2bb8841dc9da5e2f15e25f23d1872e6f
Change image load check strategy

Using the "complete" property worked for the file page, but not
for media viewer because change the src of an img doesn't reset
the prop.

This introduces a custom event that cucumber can listen to in order
to know when the full resolution image had been loaded in media
viewer.

Change-Id: I40875166e70badbb35106c4a3536c706a7c815b4