Updated mediawiki/extensions Project: mediawiki/extensions/CirrusSearch  ac86a8f64c4eb8b9b0ad836074924b6a1043ef11
Implement workItermCount for mass index job

The mass index job deserves a useful workItemCount as well.  Other jobs
are either single work item or quite rare so they don't need it.

Also, fix name of LinksUpdateSecondaryTest.  It is already in the Job
namespace so no need for Job in the name.

Change-Id: I6e511e0bb96a3f29547a9d2de097b913d436f9cf