Updated mediawiki/extensions Project: mediawiki/extensions/CentralNotice  37e512aa955cdea806b1bd6f908dc0f803f29036
Show a friendly error page when requested banner does not exist

The CentralNotice admin interface parses URL path components after the
"Special:CentralNoticeBanners/" as designating an action and a banner. If the
banner specified by the URL does not exist, CentralNotice currently throws an
exception. With this patch, a friendly error page is shown instead.

Bug: 54180
Change-Id: I4d61c700ced1001a380241d657ae0341f09ce371