Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  1c2478935fc2ffb58b87a50c78657021c5f7fd9f
Fix OPT_INDEX_TAGS not properly delegated

The options object did not do anything in this class. Now it caused
a bug. The raw mode is set to false while the constructor is called
and later changes to true.

This is the easiest way of fixing this. The underlying problem may
need more investigation.

Change-Id: I2fa8c686bd67c983fde75a91b1fa0fa2bbced6c8