Updated mediawiki/extensions Project: mediawiki/extensions/ProofreadPage  dbe5779009313698eb184df2f64a5973cf1a9427
Run the setup whenever WikiEditor is enabled

Currently the preference "showtoolbar" has no effect over WikiEditor's toolbar
(in a default installation of MediaWiki + WikiEditor).

This change makes ProofreadPage extension consistent with that by running its
setupWikiEditor in all cases where WikiEditor's toolbar is enabled (including
the case where "usebetatoolbar" is enabled but "showtoolbar" is disabled).

If/when the behavior of these two preferences changes (bug 30795), this should
be revisited to make it consistent with the new implementation.

This is a follow-up to Ic9d638f29dcd8fb9c56c04b08e4ed27fb5f31f43.

Change-Id: Ie455009df54deecc67707779e9861daf633fc690