Updated mediawiki/extensions Project: mediawiki/extensions/Collection  33be4d860dbaf491e80a758bf28c24dd8f9c72fa
Provide parsoid information to OCG if visual editor is installed.

The previous guard tested for the presence of [[Extension:Parsoid]],
which is not used outside of WMF installs, since its only job is to
manage the cache of parsoid output.  The variables referenced are
actually defined/used by [[Extension:VisualEditor]].

Change-Id: Ie1dbac73d851872a88e165bce0c2ee1ad83489c3