Updated mediawiki/extensions Project: mediawiki/extensions/CentralAuth  c9832a4268c5b784cfb2f0f259d9c689120e5149
Add script to make a list of users who will be renamed

Creates a new table, `users_to_rename`, which will be used
to store users who will eventually be renamed. It also includes
a `utf_status` column which can be used to store user's notification
status (notified by email, on-wiki, etc.), and finally to record
once they've been renamed.

A maintenance script is provided to populate the table, but hooks
and things need to be added to keep it up to date.

Change-Id: I1ba925221011dc1a082f12ad596904c673de49c3