Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  4c75f422f52349a849112c2096308e5b11688991
Update FormatAutocomments hook for I36c3a9e5

Change I36c3a9e5 to MediaWiki core changes the $pre and $post parameters
passed to the FormatAutocomments hook from strings (which must be
prepended and appended to $comment) to booleans (which must not).

For BC, this change accepts either strings or booleans and does the
right thing.

Change-Id: I6f01771e8fd6b0fcda2cfd58fcaae334c4c075d4