Updated mediawiki/extensions Project: mediawiki/extensions/Flow  4cb2d758717dd0a0ac0e2a54c48968ad512785db
Don't show 'lock' for posts; only topics

We apparently had some old code in flow_post_actions to generate some
(unfinished) "lock" link if action lock existed. "Lock" action now exists and
is showing TODO-lock for posts.
Posts shouldn't even have a lock option - that's only for topics. The link that
was generated was valid, but it doesn't make sense to have this for posts, as
the lock action only applies to the topic.

The topic title lock action link is in flow_topic_titlebar.handlebars, so will
still be present.

Change-Id: Ib3f28ac8c2e917b8869912791cebe108ccfbea21