Updated mediawiki/extensions Project: mediawiki/extensions/CentralAuth  7009430fcf293f5eda20a2103f5610f0aad3216b
SECURITY: Recheck home wiki when user merges accounts (1.24wmf22)

We check that the user can authenticate to the chosen home wiki when
doing the migration dry-run, but don't recheck when doing the actual
migration.

Since storeAndMigrate is used by migration scripts too, add a flag to
the function to signal when we need to recheck owndership.

Bug: 70468
Change-Id: I3574abe7d1f2ccb6d1241e07c900defde4acde8e