Updated mediawiki/extensions Project: mediawiki/extensions/BlueSpiceExtensions  31345ee7b01032a5eef04a34c33fc86c561ccebf
Fixed undefined object when no user is logged on read protected wikis

When not logged in on a fresh installation and the wiki is read protected,
no widgets are displayed. This lead to an undefined object. The error is
now fixed. New return is a view object with no items in it.

Change-Id: I65cae0659406f80c6ab5a54f65babfc64da52675