Updated mediawiki/extensions Project: mediawiki/extensions/BlogPage  c48035c40e26dfe786cb00b800e9b0660732cbc7
"Fixing" the by line and some other stuff

Due to the i18n changes done in Ic60108036805dd203f51a6d291957fa1323fb3b0
the "by <author name or authors' names>" line wasn't properly showing up
below the blog page title and the "About the Author" box was showing up as
"About the Authors" even when a page had only one author.

This change is more like a hotfix for the immediate problem and is likely
very English-specific. A proper fix would be nice to have.

Change-Id: Ic3933986f2eff3d5516c3c18c7ee1a350351a74f