Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  a09d648b1ed2df5192975256e60a783358883335
Initialize toolbars on existing DOM

Initializing the toolbars on the static DOM rendered by the back-end does not yet involve
the inner toolbar structure. (Still, the buttons will be removed and re-added to the DOM.)
Along with this change, the sitelinkgroup "add" buttons available in non-JS are removed in
order to receive a matching toolbar structure between JS and non-JS again.

Change-Id: I859dcc21cd17bffbb7fb265603d4d77b5fd74cf2