Updated mediawiki/extensions Project: mediawiki/extensions/Flow  f128a5663ab18a0c4fb323eb7978f7106af9863b
Show error message at correct location

showError was called on apiTarget. For replies, this target used to
be the form (IIRC), but is now the topic (which actually makes
perfect sense for what apiTarget is used for)

However, this resulted in the error being displayed in the first
error div: the one in the topic (instead of the one near the form
where the comment is being added)

Let's just find the error div closest to whatever fired the event.

Change-Id: I04f09cb470a977ae3a32157ae90aad0f2ce44351