Updated mediawiki/extensions Project: mediawiki/extensions/UrlShortener  87172b60b44748fcb610d681a766500921932d16
Use a larger alphabet for shortcodes

Allow a customisable alphabet instead of using base_convert(). By
default, use one with 64 characters. The characters were selected to be
easy to transcribe by hand, with "1", "l" and "I" omitted.

If there are any existing installations of UrlShortener, backwards
compatibility with existing URLs can be achieved with

$wgUrlShortenerIdSet = '0123456789abcdefghijklmnopqrstuvwxyz';

Change-Id: I58b53715d0822be1bf67633bb156b24ce53ba26c