Updated mediawiki/extensions Project: mediawiki/extensions/CentralAuth  58c2fc8387f0a14d266ebd0e23eb874ce9c71f14
Don't try to update a user's password if they don't exist

CentralAuthUser::matchHash() is called when we're trying to check
a local user's password during migration, so it's useless to try
and update their password. Also add an extra safeguard in
CentralAuthUser::setPassword() and debug log.

This is likely what caused bug 71223 due to the hundreds
of useless write queries that were made when trying to login.

Change-Id: I7a22fa5825e2831d4c43f85094121ec56e7cf290