Updated mediawiki/extensions Project: mediawiki/extensions/Flow  860e7bed5a518760ea8e97388b88ee3aab967b79
Hygiene: Simplify Pager

This pager needs to be updated to handle new requirements for filtering,
to do that I first needed to clean it up. This patch consolidates all of the
argument setting into the constructor so we have an easy to inspect and
understand state once the constructor has completed.

Change-Id: I4e70c65d2d75d99a4a0581488fea8296c3f81d3d