Updated mediawiki/extensions Project: mediawiki/extensions/BlogPage  0e200bbe1dfc3fe2befc3cbff767b6a3748ec059
Fix the category cloud again so that it properly ignores "Articles by User X" categories

This is likely a half-assed fix, and likely English-specific, but the
entire logic related to the 'blog-by-user-category' message and its
various use cases is broken, not i18n-compatible and needs to be rethought
from scratch. But until that, this'll have to do.

Change-Id: Ib3dd6299bd9d29b88c573d0b5680510d64b8d3cc