Updated mediawiki/extensions Project: mediawiki/extensions/Flow  aa75301c09824f86191a4c243024b166f11f09f3
Dont add permalink to topic in new-topic notification

https://trello.com/c/CoHbElok/647-single-new-topic-notification-should-link-to-board-page-when-you-click-on-link

Essentially we have the primary link of this notification already
pointing to the board page, but there is a secondary link inside the
i18n which takes users to the "wrong" place.

This patch removes that link leaving only the notification primary link.

Change-Id: I69cea4bb1a11dfaa01abc25a06a65d876ab10413