Updated mediawiki/extensions Project: mediawiki/extensions/NavigationTiming  6cbf21338846c5fa6966339c925120a48527e843
Route `$wgPercentHHVM` of users to HHVM cluster, if set

It's a bit awkward for this to go in the NavigationTiming extension, but it's a
repository I manage and its payload is loaded by default for anonymous users,
so it's a convenient way to shoehorn some JavaScript keycode into the default
JavaScript for anonymous users.

If $wgPercentHHVM is set, it is interpreted as expressing the percent chance
that a user should get the HHVM cookie.

Change-Id: I625d48f35c1059e863d634bfdc2dafa00b918c5b