Updated mediawiki/extensions Project: mediawiki/extensions/Flow  f431397e2fe02382979cc412e0a5242df9eb78ef
New-topic recentchanges must use the board title

Currently watchlisting a board does not give you entries
for new topics on that board, they are being inserted against
the topic.

The patch adds a new option, 'rc_title', that allows the action
configuration to toggle between using the owner or article title
as the recent change source.

Change-Id: Id60c6d7b05650ac12567626083814bf6526575c1