Updated mediawiki/extensions Project: mediawiki/extensions/WikimediaEvents  2baa9acb239de88782adc74676842b3bd11706b6
Groundwork for HHVM productivity analysis

For accounts created after midnight, Oct 8 UTC, set or unset the HHVM cookie
based on the user's ID: setting it if even, unsetting it if odd. Additionally,
set a config var, $wgHHVMStart, that specifies the cutoff date, so we can
compare that against wgUserRegistration in front-end logging code.

Change-Id: Ied71b503242a60f568b94f946273faece5a794bf