Updated mediawiki/extensions Project: mediawiki/extensions/Flow  116532a0f1fb9a19937197465287d5098d67bbad
Hygiene: Clean up undefined method static inspection warnings

There were approximatly 113 undefined method warnings, mostly stemming
from direct use of the Container object.  The original intention was for
Container not to be accessed directly and for all these types to be
checked in the constructor, but we have fallen off that train atm.

This patch adjusts many different uses of the container to put the value
being fetched into a variable and annotate that variable with its type.

There are still ~60 more undefined method warnings, but this patch is a good
start to cleaning them up.

Change-Id: I0d75c84d45ff99945f4c2a0a1df7834f0bda83cc