Updated mediawiki/extensions Project: mediawiki/extensions/Flow  722cb182534dbc9a537a3593e7640a5cc0494696
Hygiene: Update phpdoc for FeatureIndex

Primarily adds documentation to FeatureIndex and the Index interface
it implements.

Made a minor tweak to the Index::compareRowToOffset method, it should
have always reqiured an array as the first parameter.  Also made a
couple @todo notes for things that look like they could be simplified.

Change-Id: Ib94a889be54cccf45def60341e31d013ce510cac