Updated mediawiki/extensions Project: mediawiki/extensions/Flow  cf07bb3a998eebbd84f185ea643d3a059f33da89
Hygiene: Remove unused workflow query parameter

We used to use these, as it was the only way to identify a topic
within a board.  Now that topics have had their own page for some
time we have no need for this distinction.

Does not refactor WorkflowLoaderFactory to stop taking this parameter,
as I think the solution there is going to be more involved. Currently
the BoardContent implementation of ContentHandler utilizes the workflowId
parameter.  Potentially we should just have a method on BoardContent that
returns a workflow loader rather than accessing the factory directly, then
from api and actions we have a direct path from the input title/revision to
Flow.

Change-Id: Ied09bcba591c4f72892abbd90fded794f39368c1