Updated mediawiki/extensions Project: mediawiki/extensions/Flow  ea19d49a136fb4d211bf750296c28677c5fdca35
Text in new-topic form remains after submitting

The form was being reset after container re-initialization was
triggered.  While initializing the editors extracted the current
content of the form, and set a timeout to initialize the textarea
itself in a few moments.  In those few moments our reset ran, but
was then undone when the editor finished initializing.

This patch swaps the order such that the form is reset before the
container is initialized.

Bug: 71776
Change-Id: Ib140e44b0f7c2511518a100bc4e0eda5fb8ab3da