Updated mediawiki/extensions Project: mediawiki/extensions/Flow  030764ef94e33b2ffc5edb05a9136dfbc3027fc7
Correct placement of Anon warning in preview

In preview the anon warning was pointing to below the buttons, as the
buttons were the old part of the form remaining.  To get everything
lining up correctly we now hide the anon warning inside the form
and inject a new anon warning with the preview.

Bug: 71283
Change-Id: Ic11ff9c1b8476d56834b014609a9171b9c50d805