Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  911f97eb087f85a69590b0fbb18f887e3af7f442
Use DispatchingEntityIdParser instead of BasicEntityIdParser in repo

Constructing a new instance of BasicEntityIdParser, then calling
BasicEntityIdParser::getBuilders each time BasicEntityIdParser::parse()
is called (total of 3314 when purging a medium-size item page)
is costly, in terms of time and memory consumption.

Instead, we should be using DispatchingEntityIdParser and
can use a single instance of it via the WikibaseRepo factory.

Change-Id: I3529daefdec950a4377f503f50f07d53a29ae3b1