Updated mediawiki/extensions Project: mediawiki/extensions/Flow  9da10257eb366d0ed596a113802e0d7f508a0379
Hygiene: Pass FormatterRow for context if possible

I ran into an issue in a followup patch writing tests for the
RevisionFormatter that it was reaching out into global state via the
Collection class to get the previous revision of the one it was formatting.

This patch allows a FormatterRow instance, which is often available, to
be passed into processParams as well so it can access the already loaded
previous revision rather than reaching out through global state.

Change-Id: I416da501eeb0d41506c0192c6df581d576fe5ffc