Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  dd4f64d23ef1ce2fa324f7e10b24dda1b939d482
Use localized (in the UI language) language names in messages

The word "native" was misleading. I turned it around. My definition
of "native" is: the name of a language in that language.

This patch changes all remaining messages like "please enter a
label in Deutsch" to "please enter a label in German".

The language selector and the language names in the term box are
untouched (still the "native" language names).

Change-Id: Icad1f9e30b61fdaf33710dc50577d8e456ad0dab