Updated mediawiki/extensions Project: mediawiki/extensions/Collection  c4fbf31b7be1db2db54f0ad1e7354a51f0fef0db
Allow use of proxies to reach Rendering API endpoints.

We need to use a proxy in order to reach the PediaPress zip_post
endpoint from the production WMF servers.  With this patch we
can configure this with:

  $wgCollectionCommandToServeURL = array(
    'zip_post' => 'http://url-downloader.wikimedia.org:8080|https://pediapress.com/wmfup/',
  );

Note that the proxy to use precedes the URL and is separated by a vertical
bar.

Change-Id: I3b24b49637b9c23a1875728b62ef530e31af74e4