Updated mediawiki/extensions Project: mediawiki/extensions/Flow  bc3f150123033186787c83dea820cd2d8120d818
Hygiene: Dont reach out to $wgTitle

Actions that trigger AbstractRevision::setContent either need
to pass ina  Title, or the global $wgTitle will be used.  The $wgTitle
is mostly a fallback and isn't intended to be used anymore.

I noticed while writing a unit test for the lqt importer that PostSummary
wasn't passing the title through, so was reaching into this global state.
Patch changes PostSummary to require the Title object and pass it through.

Additionally update AbstractRevision::setNextContent and
AbstractRevision::newNextRevision to accept a title and pass it through
where applicable.

Change-Id: I6ea75b5bd415eca00ed92fe02c883abf98b39512