Updated mediawiki/extensions Project: mediawiki/extensions/BlueSpiceExtensions  5d9b7d8be6cc5f32275f1c37824e1d11e821f6ac
PermissionManager-Bugfixes

* Simplifying of TemplateEditor.saveTemplate
* Fixed: Grid gets not updated after template edit
* Fixed: Grid gets not updated after template removal
* Fixed: Tree in template editor gets not updated after template removal
* Fixed performance lag on saving settings with many namespaces shown in the grid
* Added backup creation of the settings file before saving the new settings

PatchSet 2:
I deactivated the possibility to collapse the right groups in the grid. This was necessary because ExtJS has a bug, which triggers actions on a row in the wrong record when any group above that row is collapsed. In ExtJS 5 this bugs seems to be eliminated but for now, that is the quick and painless solution we have.

Change-Id: Ib225511e28ace1bd70e6a27b33ff73f8c1aa7ed7