Updated mediawiki/extensions Project: mediawiki/extensions/Flow  0c134093907ee9f61f1cd45c23fedc6d3efea5a5
Hygiene: Make ObjectManager instantiation readable

The splits up the instantiation of the object manager such that each
piece gets its own name, I find that this makes it much more obvious
what are the arguments and what is just duplication between multiple
instantiations.

A future refactor can come through and try and remove some of the
duplication now that it is easier to see.

Change-Id: I7c4a67e136f7a5c7071ebf5f575ce637fda8fbaa