Updated mediawiki/extensions Project: mediawiki/extensions/Flow  2a0f31a01906e1996a47a13a0c0f7f9564573f8b
Make sure isRecordHidden doesn't return true is run more than once

If called more than once, the original revision that was allowed to be displayed
(the first in that workflow) would no longer return false, and as a result not
be displayed.
We loop through this method twice, once when batchloading all data, then once
per row when rendering. We want both times to returns the same result, so let's
check the exact revision that was considered visible.

Change-Id: I454a2b0993681e15144f926836df930fb63181b6