Updated mediawiki/extensions Project: mediawiki/extensions/Flow  bbd92c0837d822a532076b3148ba4970f9592d2b
Fix incorrect collapsed form state when topic sort changes.

Currently, a manual hideForm step is needed to sync up the initial
state after the template does the rendering.  This is called by
makeContentInteractive.

This was previously called by FlowBoardComponent.prototype.reinitializeBoard
(the equivalent of FlowBoardComponent.prototype.reinitializeContainer),
so this adds it back.

Bug: 72602
Change-Id: Ieb481e4b9290ce1b65cebd57d2c7910faaa8e2d8