Updated mediawiki/extensions Project: mediawiki/extensions/MultimediaViewer  cecf17319dd27fda2d9cf1867d27c83ada986b04
Fix open panel detection when scrolling

With touch devices, scrolling may get negative values (e.g., when
scrolling beyond the page limits). That kind fo scroll was triggering
the aplication of the "panel-oppen" class producing unintended UI changes.
With this patchset the class is only applied when the panel is really opened.

Change-Id: I183d8c725373ab29802650b53e26fd4820dbf499