Updated mediawiki/extensions Project: mediawiki/extensions/Phalanx  c07dff6dc6db3adab5ec829fc4f7edbb1b840e6f
i18n is a thing and not taking it into account can lead into "fun" bugs

"MediaWiki" (as in the namespace name) is not necessarily the same in
foreign languages (like in Finnish the equivalent translation is
Järjestelmäviesti, "system message").
Hence trying to whitelist blocked URLs or whatever on a wiki where the content language's
translation for NS_MEDIAWIKI was something else than "MediaWiki" would
actually trigger Phalanx when it shouldn't have done that.

Change-Id: Iaf54a4c0d90f1b0c38331b9c27d388d194c15234