Updated mediawiki/extensions Project: mediawiki/extensions/BlueSpiceExtensions  94f8d88a9d1e855b2adc37c961ca885357eb01c6
Fixed MaximizablePanel in Chrome

THIS IS A MANUAL MERGE FROM REL1_22
https://gerrit.wikimedia.org/r/#/c/169338/

In Chrome the contents of the Ext.Window derived panel where overlapped
after a 'maximize' command. One had to click on the hidden elements to
reveal them.

The panel no longer gets rendered to the element provided in the
'renderTo' config, but always to the body element. All size and position
operations are still made on 'renderTo'.

Also added logic to update absolute position on window resize

Change-Id: I3e7f46871051ed11bb8131498dce954148e46f18