Updated mediawiki/extensions Project: mediawiki/extensions/ContentTranslation  c8964d2d5510542aae610f99c2d112c33b8b0fdf
If template configuration not present, filter all

Not only at 404, but with any kind of reason, if this configuration
is not loaded, filter all.

Without this, in our cluster, for articles like
http://es.wikipedia.beta.wmflabs.org/wiki/Especial:ContentTranslation?page=Fuego&from=es&to=pt
templates will not be filtered and will clutter the UI.

We are not getting 404 when the required es-pt.json is present, but we are getting
CORS error from http://bits.beta.wmflabs.org when file is not present.

I prefer unbroken UI instead of a console.log-It does not help.

As a long term solution, we need to fix Bug 67457

Change-Id: I8a2ebe9ad1b6205192fbf16be5e173271f5d26e9