Updated mediawiki/extensions Project: mediawiki/extensions/Wikibase  5b9c2602fd4b7e7e6acc35f96db1f64815f4d46a
Simplify remoteExtPath regex

The complexity of this regular expression is just not needed. Just
search for the first occurrence of `/vendor/` or `/extensions/`
and match everything from there, that's it.

Using a regex character (`+`) as delimiter does have the advantage
that it must not be repeated when calling `preg_quote`, it's always
quoted.

Change-Id: Icbcc490879779bd6de64856411fcd862454347d4