Updated mediawiki/extensions Project: mediawiki/extensions/MobileFrontend  8278bbca4b89d73f6ebe661492f7e5dbeec80f70
Creative MaxSem friendly destruction: Wikidata Infoboxes in alpha

This replaces infoboxes in alpha of the site with a Wikidata generated one.
The idea is we can use this to highlight missing data and use it as a basis
for wikidata games.

This is rough around the edges and needs lots of design love
abut I would say is alpha ready.

Example urls:
* http://localhost:8080/wiki/Albert_Einstein?wikidataid=Q937&mobileaction=alpha
* http://localhost:8080/wiki/Barack_Obama?mobileaction=alpha&wikidataid=Q76
* http://localhost:8080/wiki/Noam%20Chomsky?wikidataid=Q9049&mobileaction=alpha

Change-Id: I824d172e30d2b88ec7101e8dfa2ce9f6719c2ddf