Updated mediawiki/extensions Project: mediawiki/extensions/FanBoxes  4ef8c81d1b3da245b860081f74175cc4b35d8888
Scoping for h1 and h3 rules

Essentially the first 187 lines should be decoupled from this file into
a separate module which is only loaded on Special:UserBoxes, but even so,
without this change there would be some skins where Special:UserBoxes
would render improperly.

SocialProfile's UserProfile (specifically UserProfilePage.php) loads this
CSS file when FanBoxes is available and enabled, so we need to limit the
scope of the h1 and h3 rules.

I chose to use both .mw-body and .mw-body-primary because at least one of
those two is present on all skins.

Change-Id: Id770404ac629b5aa26bf56e5370a7301612acaae