Updated mediawiki/extensions Project: mediawiki/extensions/SyntaxHighlight_GeSHi  966ed89270de78ee7e606db2b9d5d730a2f46297
Set Geshi version in main php file

The version was set from the ExtensionTypes hook (which runs only on
Special:Version). WikiApiary and other API consumers were unable to
detect the version.

This is an amended resubmission of the reverted d69ae1f3ac, in which
the constant was declared twice.

Change include_once to require_once for langs file (follows-up 168e1296).

Bug: T75666
Change-Id: I836e0df942a066d80255c1b68472e7ee58124357